use lazy_model and model for ui and cli.
ui 追求响应速度和加载速度, 而 cli 追求执行效率
